Clean up web api namespaces after move
